Revert "Bring 1.1 RPG - inscope.me - into ACORD"
