--- .gitignore
@@ -1,2 +1,0 @@
-
-Manacaster/Manacaster/Temp/UnityLockfile
