--- Manacaster/Manacaster/Assets/Scripts/Character.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ff6a3c90a42ccee40b326f1de69ad3ba
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
