--- Manacaster/My Scripts - Comments/Character.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: f797cfec333c2e84484c77b9c24a4c9a
-timeCreated: 1504684725
-licenseType: Pro
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
