--- Manacaster/My Scripts - Comments/Player.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: c8c88f221f6859b4896c2e4e31d3ebc4
-timeCreated: 1504628704
-licenseType: Pro
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
