+++ README.md
@@ -1,77 +1,9 @@
-# IrisDroidology/Manacaster
-
-Made by Liam Arbuckle, Manacaster is a Role Playing Game that pits you, a mad scientist, against a series of seemingly never-ending enemies in your quest to find the secret of hyperspace travel. It is loosely tied into my novel [`Star Sailors`](http://github.com/acord-robotics/stellarios/tree/starsailors), being featured in the novel. It is currently undergoing development.
-
 # Gist
-
 [Gizmotron's Gist](https://gist.github.com/IrisDroidology/897267ae735168271c8e01be28ca8d1f)
 
-
-
-## This Repository 
-
-How this repository works:
-
-* Each little bit I do I create a new branch. This way we have a consistent backup for each part of the project, in case something destructive happens
-* When I confirm that the additions work, I create a pull request and delete the arbitrary branch. The file changes, however, can still be seen in the pull request file.
-
-
-
-### Open Pull-Requests
-
-(Good as of Jan 3 2020)
-
-
-
-(None)
-
-
-
-###Closed Pull-Requests
-
-(Good as of Jan 3 2020)
-
-
-
-(None)
-
-
-
 # Log
-* [Notion Log](https://www.notion.so/gizmotron/inScope-RPG-da86aa38ed2d4bfba207262635e954aa)
+[Notion Log](https://www.notion.so/gizmotron/inScope-RPG-da86aa38ed2d4bfba207262635e954aa)
+[Issue #52](http://github.com/acord-robotics/stellarios/issues/52)
 
-* [Issue #52](http://github.com/acord-robotics/stellarios/issues/52) - [Github/Acord-Robotics](https://github.com/acord-robotics) 
-* [Issue #1](http://github.com/irisdroidology/manacaster/issues/1) - on [IrisDroidology/Manacaster](http://github.com/irisdroidology)
-
-## January 2 2020
+## January 2 2019
 I've decided to restart the tutorial. I'm going to go back to previous videos a lot, just so that I don't lose what I've learnt. I learnt a lot during this first run but I feel that I should go through it again.
-
-
-
-## Contributing
-
-Contributing to this project is easy. Just fork this repository and create a pull request when you're done. This project is completely open-source and I welcome new contributors. Consider [Joining `ACORD`](https://discord.gg/xXdJah4) to aid more in the development of our software and hardware.
-
-
-
-While this game is free, a donation is optional to keep myself fuelled with water and donuts as I create games for you. If you'd like to donate, please go to [my `website`] (http://larbuckle.tech) and help a developer in need! Donating is completely optional.
-
-
-
-# Credits
-
-Lead Developer: Liam Arbuckle 
-
-Game Art: Liam Arbuckle
-
-Base Game: InScope Studios & Kenneth Anderson
-
-
-
-Copyright LArbuckle.tech 2019-20
-
-
-
-Check out my website at http://larbuckle.tech, and my blog at http://blog.larbuckle.tech
-
-Check out ACORD, my development company, at http://acord.tech/portal
